Renamed ZBuffer library to EmuEngine
