+++ GraphicsProject.Tests/GraphicsProject.Tests.csproj
@@ -61,9 +61,9 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\ZBuffer\GraphicsProject.ZBuffer.csproj">
+    <ProjectReference Include="..\ZBuffer\GraphicsProject.EmuEngine.csproj">
       <Project>{acda8e75-8a84-4e3f-a46e-ca327f47312f}</Project>
-      <Name>GraphicsProject.ZBuffer</Name>
+      <Name>GraphicsProject.EmuEngine</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets" Condition="Exists('$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets')" />
