+++ GraphicsProject.Tests/ZBufferTests/ShapesTests/MBoxTests.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using ZBuffer.Shapes;
+using EmuEngine.Shapes;
 using System.Linq;
 using KellermanSoftware.CompareNetObjects;
 using System.Collections.Generic;
