+++ ZBuffer/Affine Transformation/IShapeEditor.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using ZBuffer.Shapes;
+using EmuEngine.Shapes;
 
-namespace ZBuffer.Affine_Transformation
+namespace EmuEngine.Affine_Transformation
 {
     interface IShapeEditor
     {
