+++ ZBuffer/Shapes/MCommonPrimitive.cs
@@ -6,7 +6,7 @@ using System.Runtime.Serialization;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ZBuffer.Shapes
+namespace EmuEngine.Shapes
 {
     [DataContract]
     public abstract class MCommonPrimitive : MShape
