+++ ZBuffer/Shapes/MShape.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Media.Media3D;
 
-namespace ZBuffer.Shapes
+namespace EmuEngine.Shapes
 {
     [DataContract]
     public abstract class MShape
