+++ ZBuffer/Tools/OrthographicCamera.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using ZBuffer.Shapes;
+using EmuEngine.Shapes;
 
-namespace ZBuffer.Tools
+namespace EmuEngine.Tools
 {
     class OrthographicCamera : Camera
     {
