+++ ZBuffer/Tools/ZBuffer.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ZBuffer.Tools
+namespace EmuEngine.Tools
 {
     class ZBuffer
     {
